feat: validate deployment-bound resource links when BPMN gets deployed
Ensure that all linked resources in a BPMN process are part of the same deployment if binding type 'deployment' is used.
If at least one resource is missing, the deployment will be rejected.
The goal is to inform users early about a missing resource and prevent incidents at runtime that would occur if a
deployment-bound link could not be resolved.

In the following cases, a linked resource is not checked:
- the id of the target resource is specified as an expression (as these can only be evaluated at runtime)
- the element that uses deployment binding is part of a non-executable process (as these will be ignored anyway)

The implementation uses an additional object to gather the model elements that use deployment binding as part of the
'createMetadata' step of the resource transformation, so that they can be validated afterward:
- this approach avoids parsing the BPMN model again to get the relevant elements
- an extra object is necessary, as the model elements cannot be stored in the DeploymentRecord or ProcessMetadata