test: don't move element outside of multi-instance
Instance migration must maintain the multi-instance state for each of
the inner instances. We cannot move an element instance out of the
multi-instance it is part of. For example, a service task `task1` which
exists within a multi-instance marked sub-process, cannot be moved to
the root-level as this would remove the multi-instance nature of this
element.