test: verify records by process instance key
While the test watcher resets the RecordingExporter, we should never
just trust it to only contain records for the process instance relevant
to this test case. Other test cases have produced other process
instances that may not have completed/terminated before this test is
run. It is possible that records for such process instances are appended
during the test, while they are not relevant.

If we don't filter out records from other process instances, we may be
checking the wrong process instance which means the test doesn't verify
what it should, but may also result in flaky test runs.