test: increase connection timeouts
Increase the socket connection timeouts to ensure that
the connection doesn't timeout early by the client.

There are multiple timeout configuration properties.
The IOReactorConfig is responsible for handling
non-blocking, i.e. async, connections. The default
connection config can also pass timeouts for the
request, as well as the socket connection.