fix: merge concurrent resolves for up to 128 hostnames
We make many requests, each of which requires a resolved hostname.
To avoid flooding the DNS server with many redundant queries, we can enable request that netty merges the requests to
resolve hostnames. A second request will wait for and use the result of the first request.

I've chosen 128 as a magic number because it's a somewhat realistic upper bound on the number of hosts we'd look up
(brokers + gateways) and is still low enough to not waste too much memory.

(cherry picked from commit e22f903434e1d678f012e889ab9ca7481dcc624c)