fix: handle output collection is null case
If the output collection is null when completing an inner instance, this
code would result in a NullPointerException. This NPE can only be thrown
when the output collection was changed to point to a different
(not-existing) variable. If the user used the SetVariable RPC, the
variable still exists. In that case, no NPE is thrown as the returned
value of getLocalVariable is a DirectBuffer to a Nil value.

A simple null check suffices to detect this problem, and we can simply
return a failure to raise an incident.