test: concurrent multi-inst activation/migration
The inner instances of a parallel multi-instance are activated using the
ProcessInstanceBatch records. This means that it's possible for a
multi-instance body to be active and migrating while its inner instances
still need to be activated.

While this isn't necessarily a very special case, the multi-instance
body processor is complex enough to warrant some additional care when
migrating them. This test case could be added without the need for
production code changes, but it may be useful to have this in the future
in case we make further changes to the multi-instance body processor. We
should never forget that instances can be migrated.