feat: add logging to rest error mappings
Ensure that the new REST API error mappings
provide logs with additional details on the
error.

Previously, a lot of information was provided
and this may pose a security risk. With the
addition of logging in the RestErrorMapper class,
more sensitive details are logged instead of being
provided in the REST API response.