feat: introduce TreePathCache
Introduce a new interface for the TreePathCache, with corresponding
implementation the FlowNodeInstanceTreePathCache.

The FlowNodeInstanceTreePathCache encapsulates the corresponding
multi-level HashMap (with SoftHashMap), and the resolvment via a
given treePathResolver.

An additional record class was created to encapsulte the composite
key consisting of: partitionId, recordKey, flowScopeKey, and
processInstanceKey. This allows us to reduce the dependency to the
Zeebe protocol. Allows us to move the cache to the common importer
module.

As the resolver is available via DI the treePath cache it can be
easily tested via unit tests. New tests and documentation were
added to cover and describe the behavior of the cache.

The new treePath cache, overcomes several  weakness of the previous
implementation (besides of not being testable).

 1. When the treePath of the flowScope was not part of the cache
    the resolver was used to resolve the treePath. This was done
    every time as the resolved treePath was not added to the cache.
    Causing unnecessary multiple calls to the resolver.

 2. The cache stored leaf flow node instances, which will never be
    requested, causing unnecesary eviction of useful data. The new
    interface allows to specify more clear when to store a treePath.

 3. The cache was a black box. It was unclear when it is accessed,
    what the results are (HIT/MISS), how big it is, etc. New metrics
    have been added to the cache implementation. Due to abstraction
    this still allows to be tested without bootstrapping unnecessary
    dependencies.