refactor: introduce FNITransformer class
Extract logic to transform flow node instance and access tree path
cache into a separate class: FNITransformer.

Creation of FNITransformer is done via lazy creation, such that we make
sure this is done only during importing, when Zeebe partitions are
available (necessary for integration tests). For lazy creation we use a
bullet-proven implementation from apache. Partitions information are
necessary to create TreePathCache which separates caches by partitions.

The extraction of the transformation code allows to better maintain and
test the logic. Multiple tests have been added with that extraction.

Furthermore, it allowed to fix several other issues/bugs.

The following issues are fixed:

 1. The treePath of a flow node instance is not updated multiple times.
    When importing completion or termination, we do not need to update
    the treePath again. With that we do not need to resolve the path.
    Especially interesting on restarts where caches might be empty, and
    importing of completion would have lead to several cache misses.
 2. We no longer store leaf flow node instances, FNI's which are not a
    container, in the cache. This reduces the potential of evicting of
    important data. This will reduce load and size of the cache.

The changes have been done on both versions of the importer 8.5 and 8.6