feat: improve java client exceptional completion
Improve the Zeebe Java client handling of failures
during REST API calls.

With the addition of REST API support for the Java client,
we added a new HttpZeebeFuture that extends CompletableFuture.
We directly use the CompletableFuture#join() there, and it
throws CompletionExceptions. In contrast, when using gRPC,
the java client calls CompletableFuture#get() instead, and it
throws ExecutionExceptions when a request fails.

We unwrap ExecutionExceptions, but don't touch CompletionExceptions.

This PR aligns the behavior and handlong of exceptional REST API
calls with gRPC calls in the Zeebe Java client.