refactor: change method signature to return a list of strings
The resource identifiers are a list of empty strings. We store this within a ResourceIdentifiers object in the state.
When we access these we can immediately return the list of strings, instead of the object. This allows us to handle the
case of no resource identifiers found within the state class, instead of everywhere we call this method.