feat: initial AuthorizationCheckBehavior
Adds a new AuthorizationCheckBehavior class. This behavior is responsible for verifying the user of the incoming
command is authorized to perform a given action. For this it takes the command, resource type, permission type and a
map of resource identifiers.
The behavior will use these parameters to fetch the permissions of this user from the state. It will compare the
resource identifiers we have there with the resource identifiers that are required. If any of them match, the user is
permitted to perform this action. It none match, the user has insufficient permissions. The command should be rejected
when this happens (not part of this commit).