feat: create AuthorizableDistributionProcessor
This processor will wrap an existing processor that requires an authorization check upon command processing.

Using the interface defined within this class it can make calls to a delegate. This delegate is responsible for the
actual command processing. The delegate will also have to provide this processor with an authorization request. This
request contains some metadata about what action the delegate is going to do. This way the
AuthorizableDistributionProcessor knows which authorization to pull from the state and can verify the permissions
accordingly.