test: set correct rest api url in tests
Set a correctly formatted REST API URL in the
Zeebe client tests. Apache HttpClient5 builds
will fail due to incorrect REST API URLs. In
the updated tests, no URI scheme was provided
resulting in incorrect parsing of the REST
API URI that leads to an exception when the
server URL for the REST API is built.