refactor: Remove reference to actor/scheduler in LogStream class
First part of https://github.com/camunda/camunda/issues/22832.
Removed reference to ActorSchedulingService from LogStream & LogStreamBuilder.

In ControlledActorClock, update() method is called when setCurrentTime is set to a positive value.
Previously, this was done by the ActorSchedulingService, but that is no longer the case.
A double check of the condition isPointInTime() is done to prevent updating the timestamp when
the timestamp is <= 0.

relates to #22832