deps: explicitly depends on protocol jackson
Without this dependency tasklist relies on zeebe-test-container
that relies on zeebe-protocol-jackson.
Because tasklist is part of the monorepo indirectly specifies 8.7.0-SNAPSHOT,
however the maven reactor doesn't understand to build this version
because zeebe-test-container is not build in this repo.
We can explicitly define the dependency to ensure that maven reactor build this as well.