fix: do not throw Exception when GCS store is not accessible on startup
If an exception is throw at startup, just log it at warning so that all
information necessary if available when reading the logs.
Because this error can be transient, it's better to continue anyway the
initialization of the broker.
A try/catch block has been added when listing the GCS bucket to report
the error with the bucket configured.

(cherry picked from commit e59ce1c717d7bcace94358e65b46f02dd798ecb2)