fix: use new record to transmit acknowledgement
The ACKNOWLEDGE command is transmitted to the other partition as a side
effect. Side effects are executed async from the command processing, so
when the side effect is eventually executed other commands may have been
processed in the meantime.

This raises a problem when the command is not copied but reused. If
another command tries to ACKNOWLEDGE a command distribution, it will
write different data into the ACKNOWLEDGE command that is scheduled for
transmission as a side effect.

We've detected some errors that seem to indicate that this has occurred
on our integration environment for a cluster that was testing the new
clock pinning endpoint.

These changes fix the issue by always constructing a new
CommandDistributionRecord for usage in the side effect for the
ACKNOWLEDGE command. This construction is costly, but necessary to
overcome this race condition.