fix: complete future with error when leadership change is cancelled
While there a transition to a leader is cancelled, the logStream will be
closed, throwing an exception when creating a Writer.

The returned future is now failed when such error happens instead of
non-completing.
Moreover, the writer is instantiated as a first step, so in case of
errors no subscriptions to the other components is done.

closes #10141

(cherry picked from commit a00267fd4241178db83f73676cf369e727637809)