feat: store commitIndex in raft MetaStore
The raft commitIndex is saved in the MetaStore, when lastFlushedIndex is
written to disk.

A new version of the SBE schema in raft-entry-schema.xml has been
created to accomodate a new field.
On startup the MetaStore reads the file and rewrites it using the
current version of the message to be able to update it in place.

closes #13790