test: fix failing test cases
The tests failed with an unauthorized rejection. This is because we now check permissions on the clock processor. Since
we only use the clock reset so we have a record to limit too, it doesn't matter if we're unauthorized. We can prevent
joining the future so we don't trigger an exception here. Because the command is still written to the log, and then
rejected it'll work fine for purpose of limiting our recording exporter.