test: test cancelling banned instances
Tests are added for cancelling banned instances. In addition to the test cases below, I did manual
testing for the following cases:

- Throw exception in the ProcessProcessor onTerminate method for the first call. Observe rejection
for the first call. Then, send cancellation command again. Expect that the banned process instance
is cancelled on the second command.
- Throw exception in the ProcessInstanceStateTransitionGuard checkStateTransition method `case:
TERMINATE_ELEMENT` for the first call. Observe rejection for the first call. Then, send
cancellation command again. Expect that the banned process instance is cancelled on the second
command.
- Create a looping process with 120000 timer instances created. Cancel the process and observe that
the cancellation works with huge amount of timers. Will therefore work for banned instances with
huge number of timer instances created. (the test setup for the support case)