test: add @ZeebeIntegration annotation
This annotation takes care of managing the TestStandaloneBroker. It also makes it so we log the records upon test
failure.

The downside of this annotation is that we now reset the RecordingExporter in a beforeEach. As a result we cannot setup
our users and clients in the beforeAll anymore. This will be fixed in the next commit.