test: extract user and client creation into the test methods themselves
Because of the refactoring of the AuthorizationsUtil, and the addition of the @ZeebeIntegration on all these test cases
we can no longer create the clients and users in the @beforeAll method. BeforeEach is also not an option, as this means
we create the users multiple times and do more work than necessary. Because of this it's been moved to the test methods
themselves.