refactor: get job from state in the JobCommandPreconditionChecker
Every single call to this check method will immediately fetch the job from the state straight after. This means in all
places where we call this, we are doing the same logic of getting the job, determining if it exists and creating a
rejection if it doesn't. We can just do it here and pass the Job along.