test: listener job vars contain completion vars
It's important that the completion operations variables are available in
the task listener's job. This allows a task listener to validate the
variables with which the user task was completed.

The variables can be made available simply in the job's variables.

As this means they merge with the process variables, the completion
operation variables hide (shadow) the process variables of the same
name.