feat: provide task completion vars to listener job
To implement this, we need to differentiate the job by it's kind. If
it's a task listener job we want to access the variables with which the
user task is being completed. We can find the variables related to the
current operation in the user task intermediate state.

Once we have both the process variables and the task variables, merging
them is difficult because they are both direct buffers containing
msgpack encoded maps. As the msgpack storage of a map differs per map
size, there's no one-size-fits-all solution to edit the buffers
in-place.

Instead, we need to decode the maps so we can merge them.

Note that we need to overwrite process variables with task variables if
they have the same name. We can easily achieve this with a putAll.

Finally, the merged map of task listener variables must be encoded to
msgpack again.