refactor: access processing state in job collector
In order to retrieve the task completion variables, we'll need to have
access to right state classes.

We'll need access to the elementInstanceState in order to find the
usertask key belonging to the task listener job.

We'll need access to the userTaskState in order to retrieve the
intermediate state of the user task. This contains the operation's
command current information like the variables of the completion.