test: ensure all processors are performing permission checks
This archunit test will ensure that all processors are checking permissions. This way devs creating processors are
forced to consider authorizations when building new features.

There's a bunch of different scenarios here:
1. A processor implements the TypedRecordProcessor interface and checks authorization directly
2. A processor implements a CommandProcessor interface and checks authorization directly
3. A processor implements a TypedRecordProcessor interface and delegates the auth check to a behaviour (eg, Job
Updating)
4. A processor implements a CommandProcessor and delegates the auth checks (eg, Job completion)
5. A processor implements the UserTaskCommandProcessor and delegates the auth checks (eg, User Task assigning)

All these scenarios are covered with 2 tests. One test to check the processors by interface, and one to verify the
delegates are performing the auth checks.