feat: backup only segment containing data not present in snapshots
Previously, all segments present in the directory would be marked
eligible for backup without coordinator with the LogCompaction.

With this change, the BackupManager will check with the Journal which
segments contain data that is not present in the selected snapshot,
limiting the possibility that these files will be canceled while a
backup is in place.

However, nothing stops the LogCompaction to delete these files if the
backup is particularly slow. To avoid such issue another synchronization
mechanism should be added to the LogCompaction.

relates #23082