feat: implement tenant cache and throttling for identity requests
Implement a tenant cache and throttling for Identity requests
in the gRPC Gateway Interceptor. The cache and throttling are
implemented through a wrapper for the identity-sdk tenants()
method.

The goal is to avoid too many concurrent requests to Identity
that may deteriorate it's service. We are currently aware that
this issue only happens with tenant-related requests. We may
expand this wrapper to other identity-sdk calls, if it is
discovered that too many concurrent calls are an issue there as
well.

This commit only adds a basic implementation to the gRPC API.
Following commits will expand this implementation with config
options. A generalization on the implementation will be applied
so that it can be re-used for the Gateway's REST API server as
well.