fix: prevent duplicate sequence numbers
Before this change we would only update the record counters on a successful flush. As a result there is a risk of a
reused sequence number when the flush fails. This happens when:

1. A record is exported
2. The client detects a flush is required immediately (synchronous flush)
3. The synchronous flush fails
4. The next flush is triggered by an asynchronous flush

If all this is true we were in a scenario where the record counters never got updated, but we did flush a record with a
sequence number to OpenSearch. A next record will receive the same sequence number in this case.

We can solve this by updating the sequence number as soon as we have written the record to the batch. Regardless of a
successful flush, the counter will be updated. If the flush still fails and the record is re-exported, the record is not
appended to the batch a second time. By using a boolean to keep track whether the exported record was appended to the
batch, we can determine if we need to update the record counters.

By using a boolean, we can easily write a unit test for this as well. If we updated the counters inside the
index method of the client, this becomes nearly impossible.

(cherry picked from commit 8324ab046b33ef470dfe95592444bfcccf81e0d8)