feat: adding new Task Listener intents
Two new intents were added:
- REJECT_TASK_LISTENER. It Represents the intent that means Task
Listener denied the operation. The creation of the next task
listener or the finalization of the original user task command
(COMPLETE) is not happening, but instead, COMPLETION_DENIED event will
be written in order to revert the User Task to CREATED state. The job
for the Task Listener itself in this case completes successfully.

- COMPLETION_DENIED.Represents the intent indicating that the User Task
will not be completed, but rather will be reverted to the CREATED state.