feat: introducing event rejection logic
`JobCompleteProcessor` will now have an additional validation to
check if JobResult `approved` field is set. If it's value is `false`,
it will write a `REJECT_TASK_LISTENER` follow up command, meaning other
task listeners (if present) won't be created and the User Task
itself won't be completed and will be reverted to `CREATED` state.

`UserTaskProcessor` will take care of processing the rejection and
writing a `COMPLETION_DENIED` follow up event