feat: do not activate jobs the user is not authorized for
Users must be authorized to activate Jobs. This authorization is done at a bpmnProcessId level. Since job activation
happens in batches, there could be multiple jobs with different process ids. Users are not necessarily authorized for
all of these jobs.
The decision was made to ignore jobs that the user lacks authorizations for. We do this by getting a set of process ids
that the user is authorized for and comparing these with the process id of the job itself. If the user is not
authorized we will skip this job and continue with the next one.
There is one exception to the rule. If the user has a wildcard permission (*) it is authorized to activate all jobs.