feat: when a follower fails to flush respond with error to leader
When a follower fails to flush, reply to the leader with a
`succeeded=false` reply, so the leader will reset the follower.
Once the follower is able to flush again, its logs will converge with
the leader's log.
The new test would not pass without the fix, because the exception
thrown when flushing would mark the follower "INACTIVE" and it would be
unable to resume processing.

A custom RaftElectionConfig is used to avoid nodes configured as faulty
to become leader.

closes #14867

(cherry picked from commit 7d1989412a717c2517d1b4b9250dc999b9a52f39)