fix: avoid reusing a persisted object for state read and write
Reusing, for example, a PersistedUser instace for reading and
writing from the DbUserState triggers a race confition where
the same memory buffers for differen PersistedUsers are used.
This leads to data inconsistencies. The solution is to pass a
different PersistedUser instance for reading values, i.e. into
the 'createColumnFamily' constructor.

The fix is applied to the DbUserState, DbGroupState,
DbTenantState, and DbGroupState, as they have similar usage
patterns for their persisted objects and reuse instances
for reading and writing."