feat: migrate compensation subscriptions
ADD COMPENSATION TO SUPPORTED EVENTS
First, compensation event type is added to supported boundary events. Later, the actual logic
implemented.

RETRIEVING RELATED SUBSCRIPTIONS
During the migration, we walk through the active elements starting from the root process instance
itself. As compensation subscriptions are created on completion of an element, the element that the
compensation subscription is created for is no longer active. That means, we won't walk through it.
But we still have access to compensation subscriptions through the process instance. For that
reason, all compensation subscriptions are handled at once for the process instance.

UNSUBSCRIBING
While unsubscribing from subscriptions that are not mapped, we add mapped subscriptions to a list
(same as what we do for other subscriptions). While doing so, we decide whether the given
subscription should be migrated or not within method `shouldCompensationBeMigrated()`. For each
subscription found, we check if compensable activity is mapped. For a task, we check if
compensation boundary events themselves is also mapped or not. This is not a must to apply the
logic, but we do this in order to align with other (message, signal, timer) boundary event
migrations. For them, you also have to map boundary events themselves.

MIGRATING
We write MIGRATED event for mapped compensation subscriptions by updating their process definition
key and compensable activity id as in target process. Later, they are updated in the state by the
applier.

SUBSCRIBING
We do not subscribe to compensation events in the target if they are not mapped. Because with the
current implementation of migration (e.g. you can only provide mapping between elements), it is not
possible to represent compensation subscriptions that are not exist in the source but should be
created in the target. E.g. if source process doesn't contain any compensation subscription and
target process has a compensation handler defined, how user can map the compensation event? It is
not possible. This logic also not supported in C7 since it is not possible. Therefore, behaviour is
aligned with C7 process instance migration as well.