feat: add processor to remove permissions
Adds the processor to remove permissions from an entity. It first checks if the user is authorized to remove
permissions. After this it checks if the entity to remove permissions from exists and only if both of these are true it
will write the event actually remove them. If not, the command will get rejected.