refactor: remove action from record
We don't need to know about the action within Zeebe. The action is used by the API to decide which intent it needs to
send to Zeebe. Because of this intent we already have context we need to know if we need to add or remove permissions.
These are separate commands. The exported events are also separate, so in the exporter it is also known what to do
without needing this property.