fix: revert common datetime format and improve deserializer
The previously defined date time format is used for writing data and reading it.
This must not change as the exported data changes compared to previous versions otherwise.

The date deserializer is made more robust so it can read the defined format but also those
produced by writers that don't use the `ConnectConfiguration` but a plain
`OffsetDatetime::toString` directly, using a colon-based offset format, potentially using 'Z'.