fix: compensation migration should work for multi instance handlers
Previously, we were receiving the compensable task by calling getElementById with the expected type.
But when getElementById is called with the expected type ExecutableActivity, it will return the
inner instance of the multi instance body, which is wrong. We wanted to retrieve the multi instance
body element itself. Later, we can retrieve its boundary events correctly.

A new method `getCompensableActivity()` is implemented to do casting manually and used in related
places.