fix: skip distributed deployments that were already created
Before storing deployments, we'd just create all resources again which works because resource creation is idempotent.
Deployment creation is not idempotent and can only be done once. To not throw an exception when the deployment is
distributed a second time, we simply check whether we already have a deployment with the same key.
If that's the case, we can directly acknowledge and skip creating the resources again