feat: introduce job result corrections
Corrections are a hardcoded object rather than a regular map as
accessing the dynamic map's entries is especially complex in our
msgpack implementation.

As each correction is optional, and our msgpack-values implementation
doesn't allow for setting null explicitly, we need a way to
differentiate set from unset. We'll do this in the same way as the user
task's changeset and changedAttributes, by introducing
correctedAttributes in addition to the corrections in a future commit,
but we already refer to it in the docs.