feat: add corrections to job result
This adds the ability to specify corrections as part of a job result.

As each correction is optional, and our msgpack-values implementation
doesn't allow for setting null explicitly, we need a way to
differentiate set from unset. We do this in the same way as the user
task's changeset and changedAttributes, by introducing
correctedAttributes in addition to the corrections.