test: ensure job corrections json serialization
This extends the existing test cases for job record serialization to
json. It shows we can serialize fully filled job results with
correctedAttributes and all corrections. It also shows we can do this
where job results are not provided.

Note that our msgpack implementation does not allow us to set the
missing values as `null`, and so entire objects are always serialized.