test: add generic test to ensure distributed commands are idempotent
This test works by running an engine with 2 partitions. It is setup to intercept the first acknowledgement from p2 to
p1. Since p1 doesn't receive the acknowledgement, it will redistribute the commmand to p2. In turn p2 wil process this
distributed command twice.

By asserting if the distribution finishes on p1, we know p2 must have resent the acknowledgement. It also means that no
unexpected error occurred when p2 reprocessed the command, so no consistency checks have failed.

It may not be 100% foolproof. P2 could still store wrong data, as long as the consistency checks don't fail. What this
test does do is trigger developer to think about the idempotency and reminds them that their processor must be
idempotent.

As a bonus, the test collects a set of processors that process distributed commands. If, at the end of the test, it
detects that we didn't run a scenario for all these processors it will fail the last scenario with the message that not
all of the processors are tested.