fix: idempotent resource deletion
Resource deletion was not idempotent. If an acknowledgement got lost, it would be retried indefinitely.
There is a catch here. Since deployments and distributions were never guaranteed to be received in order it could occur
that due to pending distribution we will receive the deletion before the deployment. This would result in a state
incosistency.