refactor: convert block comments to Javadocs for `onFinalizeCommand` methods
This commit converts the block comments above the `onFinalizeCommand` methods of `UserTaskAssignProcessor` and
`UserTaskClaimProcessor` to Javadoc with `{@inheritDoc}` tag. This ensures that the method description from the parent
interface is inherited and the IDE display the notes correctlty. Also, leveraged new markdown syntax support to format
the notes for better readability and usability within the IDE.