feat: apply user task corrected event to state
Introduces the UserTaskCorrectedApplier which stores the corrections in
the state as the intermediate state of a user task.

It simply looks up the current intermediate state and modifies the
correctable values:
- assignee
- candidateGroups
- candidateUsers
- dueDate
- followUpDate
- priority

Note that the updateIntermediateState method may provide `null` if there
is no intermediate state known for the user task. However, the corrected
event is expected to always be applied in between transitioning states
like CREATING/CREATED, ASSIGNING/ASSIGNED, UPDATING/UPDATED,
CANCELING/CANCELED, COMPLETING/COMPLETED. If it would be tried outside
of these transitioning states, the application results in a NPE which
would provide an early detection that the event cannot be applied at
that moment unless we make changes to support this. I believe this is
acceptable for now.

The test cases assume that the intermediate state is known. For the
intents CREATING and UPDATING this is not yet the case, so we simply
skip the tests. When we add support for task listeners of those
transitions, we'll add the persistence of the intermediate state, and
thus automatically enable these test cases. This allows us to ensure
that the corrected event can be applied for all expected cases.

One last thing to note is that we don't store the changedAttributes in
the persisted intermediate state. This value only makes sense for the
events themselves for communicating to consumers of the events. If the
assignee was changed in the first correction, and a second correction is
applied changing the priority only, then it would be strange that the
event would state that both the assignee and priority changed in this
correction.