refactor: moved [Elastic,Open]Search implementation of BackupRepository to webapps-backups
mplementation of BackupRepository for ElasticSearch & OpenSearch are
moved into webapp-backups.
These implementation were throwing some exceptions defined in project
operate, so their counterparts have been created in the webapp-backup
module.
A decorator BackupRepositoryWrapper to convert the generic exceptions to
the ones expected by other components in operate have been added.

The diff of this commit is quite large because a lot of code has been
moved around. The following are the points that have been slightly
modified:
- OpenSearch metadata are written in a Map<String, JsonData> and I found
in testing it the component (without mocking jackson converter) that it
would fail, so I added some logic to deserialize this Map into a
MetadataRecord. The implementation is quite basic, but it's also a
really small class.
- OpenSearch implementation would use a RichOpenSearch clients. With a
small number of changes I was able to not rely on this client given that
we are only using a small subsets of its methods.
- The logic of the snapshot naming has been abstract into a
SnapshotNameProvider class so that operate & tasklist can provide their
implementation.
- I removed the Metadata class defined in operate as it was now just a
POJO with the same fields present in webapp-backups class.

relates #24869