test: fix tests to await `ASSIGNED` event before asserting user task record
Recently introduced changes modified the behavior of user task assignment when a UserTask was created with
a defined `assignee` property. Instead of assigning the user task before the `CREATED` event, an `ASSIGNING` event
is now emitted, and the assignment is done during the `ASSIGN` (ASSIGNING/ASSIGNED) event.

This caused some tests to fail as they were asserting the `assignee` property after the `CREATED` event.
In this commit, the broken tests are updated to await the `ASSIGNED` event and assert the `assignee` property after
the user task is fully assigned.