perf: reduce test execution time in `NativeUserTaskTest` class
During fixing broken tests in the previous commit, it was observed that some tests in the `NativeUserTaskTest` class
took up to 5 seconds to execute.

The total execution time for the entire class was reduced from ~27 to ~1.3 seconds by adding a `limit` condition
to the `RecordingExporter.userTaskRecords()` stream, ensuring it stops once the last expected event is found.