test: mappings applied before updating multi instance output collection
A test is added to verify an output mapping is applied correctly when it refers to the output
collection of multi-instance body. Please note that we verify it 4 times since the fix applied will
apply output mappings once more after evaluating the output collection. See the next commit for the
fix.