feat: mappings applied before updating multi instance output collection
Existing implementation for multi instance body runs with the following steps for child instance
completion:
- Apply output mappings in child element processor onComplete()
- Update output collection in MultiInstanceBodyProcessor beforeExecutionPathCompleted()

That means, we ignore the result of updated output collection variable while applying output
mappings. As a fix, the output mapping is applied for the child instance right after evaluating the
updating the output collection.