refactor: inject pending message sub state
The transient pending message subscription state is needed in the catch
event behavior as it will need to add a side-effect to store the pending
subscription change (DELETE) when it sends it out on unsubscribing from
message catch events.

This will be needed to resolve a critical bug, where the transient state
adjusted without being rolled back when the command is rejected due to
an exception.

Co-authored-by: berkaycanbc <berkay.can@camunda.com>