fix: update transient message state as side-effect
This resolves a critical bug, where a process instance is unsubscribed
from a message catch event during instance migration even though the
migration was rejected by throwing an exception.

When the exception is thrown, we expect all state changes to be
rolled back. However, the transient state for pending process message
subscriptions was not rolled back. This led to the
PendingSubscriptionChecker to visit this pending subscription, even
though the rest of the engine considered it to be rolled back.

The transient state was updated by the DbProcessMessageSubscriptionState
on updateToClosingState which in turn is called by the
ProcessMessageSubscriptionDeletingApplier.

By moving the transient state update from the dbstate (applier) to a
side-effect in the CatchEventBehavior, we're able to ensure that the
state change only takes effect after the command was successfully
processed (completely).

It's alright to make an event applier change here without needing a new
version of the event applier, because the
DbProcessMessageSubscriptionState.onRecovered method already rebuilds
the transient state from the persisted state. This means that after
replay, the checker will be able to find the same pending subscriptions
as before our change.

We considered whether it's alright to make the change as a side-effect,
which means that the stream processor is not able to see this take
effect until the command is fully processed successfully. It is not able
to access the data in the meantime. However, this is alright because the
streamprocessor does not need to access this data. Only the checker uses
this data.

We've determined the lastSentTime right after sending the command the
first time, rather than in the side-effect, as this should be more
correct (closer in time to the actual first time sending).

Co-authored-by: berkaycanbc <berkay.can@camunda.com>